requirements.txt

numpy
opencv-contrib-python
pillow
matplotlib
tqdm
ultralytics
